+++ projects/igniteui-angular/src/lib/directives/tooltip/tooltip-target.directive.ts
@@ -86,7 +86,7 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
 
 
     /**
-     * Where to place the tooltip relative to the target element. Default value is `top`.
+     * Where to place the tooltip relative to the target element. Default value is `bottom`.
      * ```html
      * <igx-icon [igxTooltipTarget]="tooltipRef" placement="bottom-start">info</igx-icon>
      * <span #tooltipRef="tooltip" igxTooltip>Hello there, I am a tooltip!</span>
@@ -120,34 +120,34 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
     }
 
     /**
-     * Controls whether the arrow element of the tooltip is rendered.
-     * Set to true to hide the arrow. Default value is `false`.
+     * Controls whether to display an arrow indicator for the tooltip.
+     * Set to true to show the arrow. Default value is `false`.
      *
      * ```typescript
      * // get
-     * let isArrowDisabled = this.tooltip.disableArrow;
+     * let isArrowDisabled = this.tooltip.hasArrow;
      * ```
      *
      * ```typescript
      * // set
-     * this.tooltip.disableArrow = false;
+     * this.tooltip.hasArrow = true;
      * ```
      *
      * ```html
      * <!--set-->
-     * <igx-icon igxTooltipTarget [disableArrow]="true" [tooltip]="'Infragistics Inc. HQ'">info</igx-icon>
+     * <igx-icon igxTooltipTarget [hasArrow]="true" [tooltip]="'Infragistics Inc. HQ'">info</igx-icon>
      * ```
      */
     @Input()
-    public set disableArrow(value: boolean) {
+    public set hasArrow(value: boolean) {
         if (this.target) {
-            this.target.arrow.style.display = value ? 'none' : '';
+            this.target.arrow.style.display = value ? '' : 'none';
         }
-        this._disableArrow = value;
+        this._hasArrow = value;
     }
 
-    public get disableArrow(): boolean {
-        return this._disableArrow;
+    public get hasArrow(): boolean {
+        return this._hasArrow;
     }
 
     /**
@@ -348,9 +348,9 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
     private _destroy$ = new Subject<void>();
     private _autoHideDelay = 180;
     private _isForceClosed = false;
-    private _disableArrow = false;
+    private _hasArrow = false;
     private _offset = 6;
-    private _placement: TooltipPlacement = TooltipPlacement.top;
+    private _placement: TooltipPlacement = TooltipPlacement.bottom;
     private _closeButtonRef?: ComponentRef<IgxTooltipCloseButtonComponent>;
     private _closeTemplate: TemplateRef<any>;
     private _sticky = false;
@@ -367,6 +367,7 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
     ) {
         super(_element, _navigationService);
     }
+
 
 
     /**
@@ -412,8 +413,8 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
      * @hidden
      */
     public ngOnChanges(changes: SimpleChanges): void {
-        if (changes['disableArrow']) {
-            this.target.arrow.style.display = changes['disableArrow'].currentValue ? 'none' : '';
+        if (changes['hasArrow']) {
+            this.target.arrow.style.display = changes['hasArrow'].currentValue ? '' : 'none';
         }
     }
 
@@ -461,7 +462,7 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
      * ```
      */
     public showTooltip() {
-        this._showTooltip(false, true);
+        this._checksBeforeShowing(() => this._showTooltip(false, true));
     }
 
     /**
@@ -494,6 +495,20 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
         }
     }
 
+    /**
+     * A guard method that performs precondition checks before showing the tooltip.
+     * It ensures that the tooltip is not disabled and not already shown in sticky mode.
+     * If all conditions pass, it executes the provided `action` callback.
+     */
+    private _checksBeforeShowing(action: () => void): void {
+        if (this.tooltipDisabled) return;
+        if (!this.target.collapsed && this.target?.tooltipTarget?.sticky) return;
+
+        this._checkOutletAndOutsideClick();
+        this._checkTooltipForMultipleTargets();
+        action();
+    }
+
     private _hideTooltip(withDelay: boolean): void {
         if (this.target.collapsed) {
             return;
@@ -520,6 +535,8 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
 
             if (showingArgs.cancel) return;
         }
+
+        this._evaluateStickyState();
 
         this.target.timeoutId = setTimeout(() => {
             // Call open() of IgxTooltipDirective
@@ -565,9 +582,11 @@ export class IgxTooltipTargetDirective extends IgxToggleActionDirective implemen
      */
     private _checkTooltipForMultipleTargets(): void {
         if (!this.target.tooltipTarget) {
+            this.hasArrow = this._hasArrow;
             this.target.tooltipTarget = this;
         }
         if (this.target.tooltipTarget !== this) {
+            this.hasArrow = this._hasArrow;
             if (this.target.tooltipTarget.sticky) {
                 this.target.tooltipTarget._removeCloseButtonFromTooltip();
             }